chore: remove unused functions and move single-use function
